// packages/localization-bridge-schema/src/linecap-schema/bridge-context-types.ts
export type PixelRegionV02 = {
  x: number;
  y: number;
  width: number;
  height: number;
};
// ...
export type RuntimeBridgeUnitRefV02 = {
  bridgeUnitId: string;
  sourceUnitKey?: string;
};

export type RuntimeArtifactRefV02 = {
  artifactId: Uuid7;
  artifactKind: RuntimeArtifactKindV02;
  uri: string;
  hash?: string;
  mediaType?: string;
  byteSize?: number;
};

export type RuntimeTraceEventV02 = {
  traceEventId: Uuid7;
  eventKind: RuntimeTraceEventKindV02;
  bridgeUnitRef: RuntimeBridgeUnitRefV02;
  frame: number;
  traceKey?: string;
  observedText?: string;
  artifactRef?: RuntimeArtifactRefV02;
};

export type RuntimeBranchOptionV02 = {
  optionId: Uuid7;
  label?: string;
  labelBridgeUnitRef?: RuntimeBridgeUnitRefV02;
  targetRouteKey?: string;
  targetBridgeUnitRef?: RuntimeBridgeUnitRefV02;
};

export type RuntimeBranchPointEventV02 = {
  branchEventId: Uuid7;
  bridgeUnitRef: RuntimeBridgeUnitRefV02;
  frame: number;
  branchPointKey?: string;
  promptText?: string;
  options: RuntimeBranchOptionV02[];
  selectedOptionId?: Uuid7;
};

export type RuntimeCaptureV02 = {
  captureId: Uuid7;
  bridgeUnitRef: RuntimeBridgeUnitRefV02;
  evidenceTier: RuntimeEvidenceTierV02;
  frame: number;
  width: number;
  height: number;
  nonZeroPixels?: number;
  region?: PixelRegionV02;
  artifactRef: RuntimeArtifactRefV02 & { artifactKind: "screenshot" };
};

export type RuntimeRecordingV02 = {
  recordingId: Uuid7;
  bridgeUnitRef: RuntimeBridgeUnitRefV02;
  evidenceTier: RuntimeEvidenceTierV02;
  startedAtFrame: number;
  frameCount: number;
  width: number;
  height: number;
  encoding: string;
  artifactRef: RuntimeArtifactRefV02 & { artifactKind: "recording" };
};

export type RuntimeApproximationV02 = {
  approximationId: Uuid7;
  approximationTier: RuntimeApproximationTierV02;
  scope: string;
  description: string;
  affectedBridgeUnitRefs: RuntimeBridgeUnitRefV02[];
  evidenceTierCeiling: RuntimeEvidenceTierV02;
};

export type RuntimeValidationFindingV02 = {
  findingId: Uuid7;
  findingKind: RuntimeValidationFindingKindV02;
  severity: TriageSeverityV02;
  bridgeUnitRef?: RuntimeBridgeUnitRefV02;
  artifactRef?: RuntimeArtifactRefV02;
  message: string;
  evidenceTier: RuntimeEvidenceTierV02;
};

export type RuntimeReferenceComparisonV02 = {
  comparisonId: Uuid7;
  comparisonKind: RuntimeReferenceComparisonKindV02;
  status: RuntimeReferenceComparisonStatusV02;
  scope: string;
  coveredBridgeUnitRefs: RuntimeBridgeUnitRefV02[];
  artifactRef: RuntimeArtifactRefV02 & { artifactKind: "reference_comparison" };
};